Creating a node structure and a BinarySearchTree class, adding default constructor, copy constructor and destructor to the class.
The copy constructor is implemented with an additional method copyTree, which returns a pointer to the copied tree;

The destructor is implemented with the additional methods clear and clearTree. Clear is only needed to call clearTree in public class space. ClearTree removes the root and all branches and leaves of the original tree.